requirements.txt
# API Clients
yfinance>=0.2.40
requests>=2.31.0
alpha-vantage>=2.3.1

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.10
alembic>=1.13.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Data Quality & Validation
pydantic>=2.10.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
